Split sentry.utils.email into files
