+++ tests/sentry/utils/email/test_backend.py
@@ -1,0 +1,17 @@
+from sentry.testutils import TestCase
+from sentry.utils.email.backend import get_mail_backend
+
+
+class MiscTestCase(TestCase):
+    def test_get_mail_backend(self):
+        with self.options({"mail.backend": "smtp"}):
+            assert get_mail_backend() == "django.core.mail.backends.smtp.EmailBackend"
+
+        with self.options({"mail.backend": "dummy"}):
+            assert get_mail_backend() == "django.core.mail.backends.dummy.EmailBackend"
+
+        with self.options({"mail.backend": "console"}):
+            assert get_mail_backend() == "django.core.mail.backends.console.EmailBackend"
+
+        with self.options({"mail.backend": "something.else"}):
+            assert get_mail_backend() == "something.else"
